DRY DefaultAssetPath()
